raft: refactoring separate token bucket from recovery_throttle
Move token bucket algorithm to separate structure in order to use it
in other places
